requirements.txt
git+https://github.com/Rapptz/discord.py@rewrite
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
chardet==4.0.0
click==8.0.3
colorama==0.4.4
discord.py==1.7.3
Flask==2.0.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
multidict==5.2.0
python-dotenv==0.19.1
typing-extensions==3.10.0.2
Werkzeug==2.0.2
yarl==1.7.2